revise fts.nav JS
what this PR accomplishes is the issue where aria-expanded was an attribute on the navigation when moving up from small screen to large screen views, where the navigation was no longer an expandable/collapsable object.

The revision does the following:

on resize up to large screen, check to see if the navigation had an aria expanded attr of true.  if so, apply a class of was-expanded to the navigation.

whether the navigation had an aria-expanded of true or false, it still needed to be removed when in desktop mode, so now the attribute is completely removed.

when moving down to small screen view, if the navigation has the class 'was-expanded' remove that class and reset the aria-expanded to true